+++ src/error.jl
@@ -24,22 +24,22 @@ function Base.showerror(io::IO, err::UnsupportedError)
 end
 
 """
-    CannotTryResetError <: Exception
+    NotAllowedError <: Exception
 
 Abstract type for error thrown when an operation is supported but cannot be
 applied in the current state of the model.
 """
-abstract type CannotTryResetError <: Exception end
+abstract type NotAllowedError <: Exception end
 
 """
-    operation_name(err::CannotTryResetError)
+    operation_name(err::NotAllowedError)
 
 Return the name of the operation throwing the error in a gerund (i.e. -ing
 form).
 """
 function operation_name end
 
-function Base.showerror(io::IO, err::CannotTryResetError)
+function Base.showerror(io::IO, err::NotAllowedError)
     print(io, typeof(err), ": ", operation_name(err),
           " cannot be performed")
     m = message(err)
@@ -54,11 +54,11 @@ function Base.showerror(io::IO, err::CannotTryResetError)
 end
 
 """
-    message(err::Union{UnsupportedError, CannotTryResetError})
+    message(err::Union{UnsupportedError, NotAllowedError})
 
 Return a `String` containing a human-friendly explanation why the operation
 is not supported/cannot be performed. It is printed in the error message if it
 is not empty. By convention, it should be stored in the `message` field, it is
 it the case, the `message` method does not have to be implemented.
 """
-message(err::Union{UnsupportedError, CannotTryResetError}) = err.message
+message(err::Union{UnsupportedError, NotAllowedError}) = err.message