+++ src/indextypes.jl
@@ -49,20 +49,20 @@ Return a `Bool` indicating whether this index refers to a valid object in the mo
 isvalid(model::ModelLike, ref::Index) = false
 
 """
-    struct CannotDelete{IndexType <: Index} <: CannotTryResetError
+    struct DeleteNotAllowed{IndexType <: Index} <: NotAllowedError
         index::IndexType
         message::String
     end
 
 An error indicating that the index `index` cannot be deleted.
 """
-struct CannotDelete{IndexType <: Index} <: CannotTryResetError
+struct DeleteNotAllowed{IndexType <: Index} <: NotAllowedError
     index::IndexType
     message::String
 end
-CannotDelete(index::Index) = CannotDelete(index, "")
+DeleteNotAllowed(index::Index) = DeleteNotAllowed(index, "")
 
-function operation_name(err::CannotDelete)
+function operation_name(err::DeleteNotAllowed)
     return "Deleting the index $(err.index)"
 end
 
@@ -71,7 +71,7 @@ end
 
 Delete the referenced object from the model.
 """
-Base.delete!(model::ModelLike, index::Index) = throw(CannotDelete(index))
+Base.delete!(model::ModelLike, index::Index) = throw(DeleteNotAllowed(index))
 
 """
     delete!{R}(model::ModelLike, indices::Vector{R<:Index})